Sort utils and histogram functions and tests
